+++ app/integrations/class-wpml.php
@@ -14,6 +14,8 @@
 
 namespace CF_Images\App\Integrations;
 
+use stdClass;
+
 if ( ! defined( 'WPINC' ) ) {
 	die;
 }
@@ -32,8 +34,9 @@ class Wpml {
 	public function __construct() {
 		add_filter( 'cf_images_media_post_id', array( $this, 'get_original_image_id' ) );
 		add_action( 'cf_images_before_wp_query', array( $this, 'remove_wpml_filters' ) );
-		add_action( 'wpml_after_duplicate_attachment', array( $this, 'ignore_attachment' ), 10, 2 );
-		add_action( 'wpml_after_copy_attached_file_postmeta', array( $this, 'ignore_attachment' ), 10, 2 );
+		add_action( 'cf_images_upload_success', array( $this, 'update_image_meta' ), 10, 2 );
+		add_action( 'cf_images_remove_success', array( $this, 'image_removed_from_cf' ) );
+		add_filter( 'cf_images_wp_query_args', array( $this, 'add_wp_query_args' ), 10, 2 );
 	}
 
 	/**
@@ -75,16 +78,85 @@ class Wpml {
 	}
 
 	/**
-	 * Fires when an attachment is duplicated.
+	 * Get translations for an image.
 	 *
-	 * Duplicated images do not need to be processed, otherwise this causes double uploads to Cloudflare.
+	 * @since 1.9.0
 	 *
-	 * @param int $attachment_id            The ID of the source/original attachment.
-	 * @param int $duplicated_attachment_id The ID of the duplicated attachment.
+	 * @param int $attachment_id Attachment ID.
 	 *
-	 * @since 1.4.0
+	 * @return array
 	 */
-	public function ignore_attachment( int $attachment_id, int $duplicated_attachment_id ) {
-		update_post_meta( $duplicated_attachment_id, '_cloudflare_image_skip', true );
+	private function get_translations( int $attachment_id ): array {
+		global $sitepress;
+
+		if ( ! $sitepress || ! method_exists( $sitepress, 'get_element_trid' ) ) {
+			return array();
+		}
+
+		$translation_id = $sitepress->get_element_trid( $attachment_id, 'post_attachment' );
+		return $sitepress->get_element_translations( $translation_id, 'post_attachment', true );
+	}
+
+	/**
+	 * Update the meta for all images.
+	 *
+	 * @since 1.9.0
+	 *
+	 * @param int      $attachment_id Original attachment ID.
+	 * @param stdClass $results       Upload results.
+	 */
+	public function update_image_meta( int $attachment_id, stdClass $results ) {
+		$translations = $this->get_translations( $attachment_id );
+
+		foreach ( $translations as $translation ) {
+			if ( $translation->original ) {
+				continue;
+			}
+
+			update_post_meta( $translation->element_id, '_cloudflare_image_id', $results->id );
+		}
+	}
+
+	/**
+	 * Remove meta from all translatable images when the main image is removed from Cloudflare.
+	 *
+	 * @since 1.9.0
+	 *
+	 * @param int $attachment_id Attachment ID.
+	 */
+	public function image_removed_from_cf( int $attachment_id ) {
+		$translations = $this->get_translations( $attachment_id );
+
+		foreach ( $translations as $translation ) {
+			if ( $translation->original ) {
+				continue;
+			}
+
+			delete_post_meta( $translation->element_id, '_cloudflare_image_id' );
+		}
+	}
+
+	/**
+	 * Adjust the WP_Query args for bulk offload action.
+	 *
+	 * @since 1.9.0
+	 * @see Ajax::get_wp_query_args()
+	 *
+	 * @param array  $args   WP_Query args.
+	 * @param string $action Executing action.
+	 *
+	 * @return array
+	 */
+	public function add_wp_query_args( array $args, string $action ): array {
+		if ( 'upload' !== $action ) {
+			return $args;
+		}
+
+		$args['meta_query'][] = array( // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_query
+			'compare' => 'NOT EXISTS',
+			'key'     => 'wpml_media_processed',
+		);
+
+		return $args;
 	}
 }