+++ app/modules/class-module.php
@@ -14,6 +14,7 @@
 
 namespace CF_Images\App\Modules;
 
+use CF_Images\App\Settings;
 use CF_Images\App\Traits\Helpers;
 
 if ( ! defined( 'WPINC' ) ) {
@@ -102,10 +103,12 @@ abstract class Module {
 	 *
 	 * @since 1.1.3
 	 *
+	 * @param int $attachment_id Optional. Attachment ID.
+	 *
 	 * @return bool
 	 */
-	protected function can_run(): bool {
-		if ( $this->is_rest_request() || wp_doing_cron() ) {
+	protected function can_run( int $attachment_id = 0 ): bool {
+		if ( $this->is_rest_request( $attachment_id ) || wp_doing_cron() ) {
 			return false;
 		}
 
@@ -122,9 +125,19 @@ abstract class Module {
 	 *
 	 * @since 1.2.0
 	 *
+	 * @param int $attachment_id Optional. Attachment ID.
+	 *
 	 * @return bool
 	 */
-	private function is_rest_request(): bool {
+	private function is_rest_request( int $attachment_id = 0 ): bool {
+		if ( $attachment_id ) {
+			// We must rely on the REST API endpoints if full offload is enabled.
+			$deleted = get_post_meta( $attachment_id, '_cloudflare_image_offloaded', true );
+			if ( $deleted && apply_filters( 'cf_images_module_enabled', false, 'full-offload' ) ) {
+				return false;
+			}
+		}
+
 		$wordpress_has_no_logic = filter_input( INPUT_GET, '_wp-find-template' );
 		$wordpress_has_no_logic = sanitize_key( $wordpress_has_no_logic );
 
@@ -163,8 +176,33 @@ abstract class Module {
 			$module = $this->module;
 		}
 
-		$settings = get_option( 'cf-images-settings', \CF_Images\App\Settings::get_defaults() );
+		$settings = apply_filters( 'cf_images_settings', get_option( 'cf-images-settings', Settings::get_defaults() ) );
 
 		return apply_filters( 'cf_images_module_status', $settings[ $module ] ?? $fallback, $module );
+	}
+
+	/**
+	 * In certain cases offloading should be disabled.
+	 *
+	 * @since 1.9.0
+	 *
+	 * @return bool
+	 */
+	public function can_offload(): bool {
+		if ( filter_input( INPUT_GET, 'cf-images-disable' ) ) {
+			return false;
+		}
+
+		// Full offload overrides all other settings, because there are no local images available.
+		if ( $this->is_module_enabled( false, 'full-offload' ) ) {
+			return true;
+		}
+
+		if ( $this->is_module_enabled( false, 'no-offload-user' ) && defined( 'LOGGED_IN_COOKIE' ) ) {
+			// Check logged-in user cookie.
+			return empty( $_COOKIE[ LOGGED_IN_COOKIE ] );
+		}
+
+		return true;
 	}
 }