+++ cf-images.php
@@ -14,7 +14,7 @@
  * Plugin Name:       Offload Media to Cloudflare Images
  * Plugin URI:        https://vcore.au
  * Description:       Offload media library images to the `Cloudflare Images` service.
- * Version:           1.8.0
+ * Version:           1.9.0
  * Author:            Anton Vanyukov
  * Author URI:        https://vcore.au
  * License:           GPL-2.0+
@@ -31,7 +31,7 @@ if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-define( 'CF_IMAGES_VERSION', '1.8.0' );
+define( 'CF_IMAGES_VERSION', '1.9.0' );
 define( 'CF_IMAGES_DIR_URL', plugin_dir_url( __FILE__ ) );
 
 require_once 'app/class-activator.php';
@@ -49,6 +49,7 @@ register_deactivation_hook( __FILE__, array( 'CF_Images\App\Activator', 'deactiv
  */
 function run_cf_images() {
 	require_once __DIR__ . '/app/traits/trait-ajax.php';
+	require_once __DIR__ . '/app/traits/trait-empty-init.php';
 	require_once __DIR__ . '/app/traits/trait-helpers.php';
 	require_once __DIR__ . '/app/traits/trait-stats.php';
 	require_once __DIR__ . '/app/class-core.php';